fix compiler warning
git-svn-id: http://svn.code.sf.net/p/tango-cs/code/bindings/PyTango/trunk@22673 4e9c00fd-8f2e-0410-aa12-93ce3db5e235

+++ src/boost/cpp/device_attribute_numpy.hpp
@@ -252,7 +252,7 @@ namespace PyDeviceAttribute {
                 }
             }
         } else {
-            for (unsigned long x=0; x < dim_x; ++x) {
+            for (Py_ssize_t x=0; x < dim_x; ++x) {
                 PyObject* dataObj = PyArray_GETITEM(array, iter->dataptr);
                 const object py_data = object( handle<>( dataObj ) );
